requirements.txt
black==21.12b0
click==8.0.3
Faker==11.3.0
Flask==2.0.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
text-unidecode==1.3
tomli==1.2.3
typing-extensions==4.0.1
Werkzeug==2.0.2
